+++ packages/terra-date-picker/src/DateInput.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import Button from 'terra-button';
 import IconCalendar from 'terra-icon/lib/icon/IconCalendar';
 import Input from 'terra-form-input';
+import { injectIntl, intlShape } from 'react-intl';
 import DateUtil from './DateUtil';
 import styles from './DatePicker.module.scss';
 
@@ -18,6 +19,11 @@ const propTypes = {
    */
   // eslint-disable-next-line react/forbid-prop-types
   inputAttributes: PropTypes.object,
+  /**
+   * @private
+   * intl object programmatically imported through injectIntl from react-intl.
+   * */
+  intl: intlShape.isRequired,
   /**
    * Name of the date input.
    */
@@ -74,15 +80,6 @@ const defaultProps = {
   value: undefined,
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-};
-
 // eslint-disable-next-line react/prefer-stateless-function
 class DatePickerInput extends React.Component {
   constructor(props) {
@@ -127,6 +124,7 @@ class DatePickerInput extends React.Component {
     const {
       buttonRefCallback,
       inputAttributes,
+      intl,
       name,
       onBlur,
       onChange,
@@ -147,8 +145,8 @@ class DatePickerInput extends React.Component {
 
     const additionalInputProps = Object.assign({}, customProps, inputAttributes);
 
-    const dateValue = DateUtil.convertToISO8601(value, DateUtil.getFormatByLocale(this.context.intl.locale));
-    const buttonText = this.context.intl.formatMessage({ id: 'Terra.datePicker.openCalendar' });
+    const dateValue = DateUtil.convertToISO8601(value, DateUtil.getFormatByLocale(intl.locale));
+    const buttonText = intl.formatMessage({ id: 'Terra.datePicker.openCalendar' });
 
     return (
       <div className={styles['custom-input']}>
@@ -191,6 +189,5 @@ class DatePickerInput extends React.Component {
 
 DatePickerInput.propTypes = propTypes;
 DatePickerInput.defaultProps = defaultProps;
-DatePickerInput.contextTypes = contextTypes;
 
-export default DatePickerInput;
+export default injectIntl(DatePickerInput);