+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,7 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
+* Updated to interface with react-intl's `FormattedMessage` rather than old React Context API.
 * When the date-time picker time clarification modal is open on touch devices, it will now focus on the first interactive DOM element within it. This enables focus to be shifted into the modal correctly when using VoiceOver on iOS.
 
 4.7.0 - (July 16, 2019)